requirements.txt
Django>=5.2
celery>=5.3
redis>=5.0
mypy>=1.8
mypy-extensions>=1.0
mypy-django-plugin>=1.3
django-stubs>=4.2
pytest>=7.4.0
pytest-django>=4.5.0
pytest-freezegun>=0.2.0
factory-boy>=3.3.0 